Refactored provider detection tests
