+++ CryptoAnalysis/src/test/java/tests/providerdetection/examples/ProviderDetectionExample12.java
@@ -1,4 +1,4 @@
-package tests.providerdetection;
+package tests.providerdetection.examples;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
@@ -6,30 +6,16 @@ import java.util.Random;
 
 import javax.crypto.KeyGenerator;
 
-public class ProviderDetectionExample14 {
+public class ProviderDetectionExample12 {
 
 	public static void main(String[] args) throws NoSuchAlgorithmException, NoSuchProviderException {
 		Random rand = new Random();
 		int n = rand.nextInt(2);
 		
 		String pString1 = "BC";
+		String pString2 = "BCPQC";
 		
-		switch(n) {
-		case 0:
-			pString1="BC";
-			break;
-		case 1: 
-			pString1="BCPQC";
-			break;
-		case 2:
-			pString1="BC";
-			break;
-		default:
-			pString1="BCPQC";
-			break;
-		}
-		
-		KeyGenerator keygenerator = KeyGenerator.getInstance("AES", pString1);
+		KeyGenerator keygenerator = KeyGenerator.getInstance("AES", ((n%2==0) ? pString1 : pString2));
 		keygenerator.generateKey();
 	}
