+++ CryptoAnalysis/src/test/java/tests/providerdetection/examples/ProviderDetectionExample13.java
@@ -1,4 +1,4 @@
-package tests.providerdetection;
+package tests.providerdetection.examples;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
@@ -6,16 +6,22 @@ import java.util.Random;
 
 import javax.crypto.KeyGenerator;
 
-public class ProviderDetectionExample12 {
+public class ProviderDetectionExample13 {
 
 	public static void main(String[] args) throws NoSuchAlgorithmException, NoSuchProviderException {
 		Random rand = new Random();
 		int n = rand.nextInt(2);
 		
 		String pString1 = "BC";
-		String pString2 = "BCPQC";
 		
-		KeyGenerator keygenerator = KeyGenerator.getInstance("AES", ((n%2==0) ? pString1 : pString2));
+		if(n%2==0) {
+			pString1 = "BC";
+		}
+		else {
+			pString1 = "BCPQC";
+		}
+		
+		KeyGenerator keygenerator = KeyGenerator.getInstance("AES", pString1);
 		keygenerator.generateKey();
 	}
