+++ CryptoAnalysis/src/test/java/tests/providerdetection/examples/ProviderDetectionExample14.java
@@ -1,4 +1,4 @@
-package tests.providerdetection;
+package tests.providerdetection.examples;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
@@ -6,7 +6,7 @@ import java.util.Random;
 
 import javax.crypto.KeyGenerator;
 
-public class ProviderDetectionExample13 {
+public class ProviderDetectionExample14 {
 
 	public static void main(String[] args) throws NoSuchAlgorithmException, NoSuchProviderException {
 		Random rand = new Random();
@@ -14,11 +14,19 @@ public class ProviderDetectionExample13 {
 		
 		String pString1 = "BC";
 		
-		if(n%2==0) {
-			pString1 = "BC";
-		}
-		else {
-			pString1 = "BCPQC";
+		switch(n) {
+		case 0:
+			pString1="BC";
+			break;
+		case 1: 
+			pString1="BCPQC";
+			break;
+		case 2:
+			pString1="BC";
+			break;
+		default:
+			pString1="BCPQC";
+			break;
 		}
 		
 		KeyGenerator keygenerator = KeyGenerator.getInstance("AES", pString1);
