+++ CryptoAnalysis/src/test/java/tests/providerdetection/examples/ProviderDetectionExample5.java
@@ -1,14 +1,14 @@
-package tests.providerdetection;
+package tests.providerdetection.examples;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
-
 import javax.crypto.KeyGenerator;
 
-public class ProviderDetectionExample6 {
+public class ProviderDetectionExample5 {
 
 	public static void main(String[] args) throws NoSuchAlgorithmException, NoSuchProviderException {
-		KeyGenerator keygenerator = KeyGenerator.getInstance("AES", "BC");
+		String p1 = "BC";
+		KeyGenerator keygenerator = KeyGenerator.getInstance("AES", p1);
 		keygenerator.generateKey();
 	}
