+++ CryptoAnalysis/src/test/java/tests/providerdetection/examples/ProviderDetectionExample6.java
@@ -1,14 +1,14 @@
-package tests.providerdetection;
+package tests.providerdetection.examples;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
+
 import javax.crypto.KeyGenerator;
 
-public class ProviderDetectionExample5 {
+public class ProviderDetectionExample6 {
 
 	public static void main(String[] args) throws NoSuchAlgorithmException, NoSuchProviderException {
-		String p1 = "BC";
-		KeyGenerator keygenerator = KeyGenerator.getInstance("AES", p1);
+		KeyGenerator keygenerator = KeyGenerator.getInstance("AES", "BC");
 		keygenerator.generateKey();
 	}
