+++ CryptoAnalysis/src/test/java/tests/providerdetection/examples/ProviderDetectionExample7.java
@@ -1,13 +1,14 @@
-package tests.providerdetection;
+package tests.providerdetection.examples;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
 
-public class ProviderDetectionExample8 {
+public class ProviderDetectionExample7 {
 
 	public static void main(String[] args) throws NoSuchAlgorithmException, NoSuchProviderException {
-		MessageDigest md = MessageDigest.getInstance("AES", "BC");
+		String p1 = "BC";
+		MessageDigest md = MessageDigest.getInstance("AES", p1);
 		byte[] input = "message".getBytes();
 		md.digest(input);
 	}
