+++ CryptoAnalysis/src/test/java/tests/providerdetection/examples/ProviderDetectionExample8.java
@@ -1,15 +1,13 @@
-package tests.providerdetection;
+package tests.providerdetection.examples;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
 
-import org.bouncycastle.jce.provider.BouncyCastleProvider;
-
-public class ProviderDetectionExample4 {
+public class ProviderDetectionExample8 {
 
 	public static void main(String[] args) throws NoSuchAlgorithmException, NoSuchProviderException {
-		MessageDigest md = MessageDigest.getInstance("AES", new BouncyCastleProvider());
+		MessageDigest md = MessageDigest.getInstance("AES", "BC");
 		byte[] input = "message".getBytes();
 		md.digest(input);
 	}
